requirements.txt
# Basic libraries for data handling and processing
numpy>=1.21.0
pandas>=1.3.0

# Machine learning library
scikit-learn>=0.24.2

# Natural language processing library
nltk>=3.6.2

# Web framework for building APIs
flask>=2.0.1

# Library for managing environment configurations
python-dotenv>=0.19.0

# Testing tools
pytest>=6.2.4

# Logging utility
loguru>=0.5.3